docker/requirements.txt
aiogram==2.25.1
Flask==2.3.3
python-dotenv==1.0.1
requests==2.31.0
mysql-connector-python==8.2.0
aiohttp==3.8.6
flake8
black
bandit
isort